+++ Logic/src/Stopwatch.cpp
@@ -16,22 +16,22 @@ logic::Stopwatch::Stopwatch() {
     t_previous = t_start;
 }
 
-double logic::Stopwatch::get_delta_time() {
+float logic::Stopwatch::get_delta_time() {
     std::chrono::system_clock::time_point now = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double, std::milli> delta = now - t_previous; // difference in milliseconds
+    std::chrono::duration<float, std::milli> delta = now - t_previous; // difference in milliseconds
     t_previous = now;
     return delta.count();
 }
 
-double logic::Stopwatch::get_start_time() {
+float logic::Stopwatch::get_start_time() {
     std::chrono::system_clock::time_point now = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double, std::milli> delta = now - t_start; // difference in milliseconds
+    std::chrono::duration<float, std::milli> delta = now - t_start; // difference in milliseconds
     return delta.count();
 }
 
-double logic::Stopwatch::get_time_between(std::chrono::system_clock::time_point time_point1,
+float logic::Stopwatch::get_time_between(std::chrono::system_clock::time_point time_point1,
                                           std::chrono::system_clock::time_point time_point2) {
-    std::chrono::duration<double, std::milli> delta = time_point1 - time_point2;
+    std::chrono::duration<float, std::milli> delta = time_point1 - time_point2;
     return delta.count();
 }
