+++ Representation/include/CoinView.h
@@ -41,7 +41,7 @@ public:
      * @brief does nothing
      * @param dt delta-time between updates
      */
-    void update(double dt) override {}
+    void update(float dt) override {}
 };
 } // namespace representation
 #endif // PACMAN_COINVIEW_H
